ByteTracker_approach/requirements.txt
# Requirements for Player Re-Identification System
# Implementation of research paper: "Player Re-Identification Using Body Part Appearances"

# Core dependencies
torch>=1.11.0
torchvision>=0.12.0
opencv-python>=4.5.0
numpy>=1.21.0
pillow>=8.0.0
tqdm>=4.62.0
pyyaml>=6.0

# YOLO and tracking
ultralytics>=8.0.0

# Optional dependencies for enhanced features
scikit-learn>=1.0.0  # For clustering and metrics
matplotlib>=3.5.0    # For visualization
seaborn>=0.11.0     # For advanced plotting

# Development dependencies (optional)
pytest>=6.0.0       # For testing
black>=22.0.0       # For code formatting
flake8>=4.0.0       # For linting

# Platform-specific (uncomment if needed)
# For CUDA support (Linux/Windows):
# torch>=1.11.0+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html
# torchvision>=0.12.0+cu113 -f https://download.pytorch.org/whl/cu113/torch_stable.html

# For Apple Silicon Macs:
# torch>=1.11.0
# torchvision>=0.12.0
gdown